Worker/Seller V.1
Worker/Seller V.1
Hopefull this works and ask me fi there is any problems getting it downloaded and working.

It has Scripts, ScriptbleObjects and Prefab for the worker AI.
They are editble for changing the ammount of Gain and Cost.
This Includes:
- Prefab/ResourceProducer.
- Resource/ResourceProductionData.
- Resource/ResourceProductionData/
Resource Data Assets (Worker and Seller.)
- Scenes/Amys_Work (This Scene should include everything of the work and how to copy it into the main.)
- Scripts/Resources/Data
- Scripts/Resources/Purchasable
- Scripts/Resources/ResourceAmount
- Scripts/Resources/Producer
- Scripts/Resources/SetUp

Ask if any questions.